requirements.dev.txt
flake8>=3.9.2,<3.10
pytest-django>=4.10.0,<5.0.0

